tickets 882 885 887 888 889 890

+++ app/controllers/Admin/AdminRegionsController.php
@@ -62,26 +62,34 @@ class AdminRegionsController extends BaseController
 	    $supplierMainRegions = SiteDetails::where('regions_id','!=','NULL')->where('regions_id','!=',0)->lists('regions_id');
 	    $bindings = array_unique(array_merge($supplierMainRegions,$bindings));
 	    $regions = Region::whereNotIn('id',$this->ids)->get();
+	    $names = array();
+	    $restricted = array();
 	    foreach ($regions as $region) {
 	    	$this->idSet($region);
 	    	$test = array_intersect($this->branchIds,$bindings);
 	    	if(count($test))
-	    	{
-	    		$region = Region::where('id','=',current($test))->first();
-	    		return Response::json(array('error'=>'לא ניתן למחוק אזור "'.$region->name.'" מכיוון שהינו משויך לאחד הספקים','tree'=>$regionsTree),501);
-	    	}
+	    		$restricted = array_merge($restricted,$this->branchIds);
 	    }
 	    // if(count($bindings))
 	    // {
 	    // 	$region = Region::where('id','=',current($bindings))->first();
 	    // 	return Response::json(array('error'=>'לא נינן למחוק אזור '.$region->name.' מכיוון שהוא משויך לאחד הספקים','tree'=>$regionsTree),501);
 	    // }
-	   
+	   	$restricted = array_unique($restricted);
+	   	foreach ($restricted as $key => $value) {
+	   		$reg = Region::find($value);
+	   		if($reg)
+	   			$names[] = $reg->name;
+	   	}
 	    foreach ($regions as $region) {
-	    	$region->delete();
+	    	if(!in_array($region->id,$restricted))
+	    		$region->delete();
 	    }
 	    $regions = Region::orderBy('name','ASC')->get();
 	    $regionsTree = json_decode($this->index()->getContent(),true);
-		return Response::json(array('regions'=>$regions,'tree'=>$regionsTree),200);
+	    if(count($names))
+	    	return Response::json(array('error'=>'לא ניתן למחוק אזור "'.implode(',',$names).'" מכיוון שהינו משויך לאחד הספקים','tree'=>$regionsTree),501);
+	    else
+			return Response::json(array('regions'=>$regions,'tree'=>$regionsTree),200);
 	}
 }
